I want to commit my latest changes because I am going to do some major client-side modifications after this commit and want to make sure that the latest client stuff is there. 1) Can't move user figures by clicking on them now and dragging them around. 2) Made some small changes in the Data Server - namely made sure that the data stored in the Data Server is case sensitive because using case-insensitive data started to present problems with some websites which rely on this kind of thing.
Am working on making more hallways i.e. if there are a lot of links and more than one hallway is needed. Will try to create a row of hallways following the one Navdeep creates. After studying and taking apart a few of Navdeep's methods I think I have a good idea of how to do this...