Probes now work!! yeah! but I will need to run some more debugging sessions.
I designed the probe manager - the thing that manages probes to use the same dynamic class loading as frax so it can reuse the protocols that frax supports. You just need to change one line in each protocol which you move over (just change the method called goPlug to goProbe and you're all set. Of course you also need to change the package name appropriately - I guess that's two lines then :) . Just wanted to keep it very easy to upgrade.

If you want to configure what data is compared between the old data and the new data you can find this information in probe.cfg. There are 3 possible things to compare: metadata, data, and length. Metadata includes date-last-modified, length, and other metadata generated by a protocol so this is protocol specific and varies with protocol. Data is the actual data of the data source - comparing this is slow and tedious ,in the worst case scenario, because the amount of data can be large. It is left up to the administrator of the probe manager to decide whether data needs to be compared. Another quick way to compare data sources is purely by length. This is the quickest way of doing this. All these options are present in probe.cfg.