Side Door functions are added.
For time being, "Side Door" is created by "Single Click"


"Side Door" should be created when the object is DOUBLE clicked and
if the object is CONTAINER

Yet, somehow, making "Double Click" is very hard. (most cases, it will be
treated as "Single Click" x 2.

So, for the time being, "Creating Side Door" is placed in HandleLeftMouseClick()

In order for "side door" to work with Double-click, move DrawSideDoor()
from HandleLeftClick() to HandleLeftDoubleClick().